+++ packages/deepyr/lib/src/base/component_stylings.dart
@@ -118,6 +118,7 @@ mixin AllComponentStylings
         RangeStyling,
         SelectStyling,
         SkeletonStyling,
+        StackStyling,
         StatsStyling,
         StatStyling,
         StatFigureStyling,